Address PR comments, experimental split
