+++ hypothesis-python/tox.ini
@@ -16,14 +16,15 @@ passenv=
     TOXENV
     # Allow CI builds (or user builds) to force coloured terminal output.
     PY_COLORS
-setenv=
+setenv =
+    # If PY...ING is set in calling environment, use that; default to 1 
     PYTHONWARNDEFAULTENCODING={env:PYTHONWARNDEFAULTENCODING:1}
     brief: HYPOTHESIS_PROFILE=speedy
 commands =
     full: bash scripts/basic-test.sh
     brief: python -bb -X dev -m pytest -n auto tests/cover/test_testdecorators.py {posargs}
     cover: python -bb -X dev -m pytest -n auto tests/cover/ tests/pytest/ tests/conjecture/ {posargs}
-    rest: python -bb -X dev -m pytest -n auto tests/ --ignore=tests/cover/ --ignore=tests/pytest/ --ignore=tests/conjecture/ --ignore=tests/nocover/ --ignore=tests/quality/ --ignore=tests/ghostwriter/ --ignore=tests/patching/ {posargs}
+    rest: python -bb -X dev -m pytest -n auto tests/ --ignore=tests/cover/ --ignore=tests/pytest/ --ignore=tests/conjecture/ --ignore=tests/nocover/ --ignore=tests/quality/ --ignore=tests/ghostwriter/ --ignore=tests/patching/ --ignore=tests/crosshair/test_crosshair.py {posargs}
     conjecture: python -bb -X dev -m pytest -n auto tests/conjecture/ {posargs}
     nocover: python -bb -X dev -m pytest -n auto tests/nocover/ {posargs}
     niche: bash scripts/other-tests.sh
